#1050: Task Manager - Old-scheduler's logic replaced by task manager's logic
Small fixes.
Isolated transaction management in a separated implementation (just for looking forward).

+++ core/src/main/java/org/cmdbuild/data/store/scheduler/SchedulerJobParameterConverter.java
@@ -23,7 +23,7 @@ public class SchedulerJobParameterConverter extends BaseStorableConverter<Schedu
 	public SchedulerJobParameter convert(final CMCard card) {
 		return SchedulerJobParameter.newInstance() //
 				.withId(card.getId()) //
-				.withOwner(card.get(SCHEDULER_ID, Long.class)) //
+				.withOwner(card.get(SCHEDULER_ID, Integer.class).longValue()) //
 				.withKey(card.get(KEY, String.class)) //
 				.withValue(card.get(VALUE, String.class)) //
 				.build();
